move the classes that is not View class into other files: - BodyCondition - NutritionPlan - PlanPeriod - UserDefineMeal
